+++ src/server/global_notifier.cpp
@@ -382,7 +382,7 @@ SharedRealm GlobalNotifier::ChangeNotification::get_new_realm() const
     return m_new_realm;
 }
 
-std::unordered_map<std::string, _impl::ObjectChangeSet> const& GlobalNotifier::ChangeNotification::get_changes() const
+std::unordered_map<std::string, ObjectChangeSet> const& GlobalNotifier::ChangeNotification::get_changes() const
 {
     if (m_have_calculated_changes)
         return m_changes;
